+++ src/Components/Table/Table.js
@@ -51,8 +51,8 @@ const TableData = styled.td`
   display: ${({ copy }) => (copy ? 'flex' : '')};
   align-items: ${({ copy }) => (copy ? 'center' : '')};
   border: none;
-  text-decoration: ${({ skipped }) => skipped && 'line-through'};
-  font-style: ${({ skipped }) => skipped && 'italic'};
+  text-decoration: ${({ skipped, scheduled }) => skipped && !scheduled && 'line-through'};
+  font-style: ${({ skipped, scheduled }) => (skipped || scheduled) && 'italic'};
   color: ${({ color }) => color && color};
 `;
 const Pagination = styled(BasePagination)`
@@ -186,10 +186,12 @@ const Table = ({
 
         const {
           link = false,
+          externalLink = '',
           value = '--',
           hover = false,
           icon,
           skipped = false,
+          scheduled = false,
           iconColor = theme.FONT_LINK,
           color = '',
           size = '1.4rem',
@@ -209,7 +211,8 @@ const Table = ({
           <TableData
             title={hover || value}
             key={displayName}
-            skipped={skipped}
+            skipped={skipped && !scheduled}
+            scheduled={scheduled}
             copy={copy || (!isEmpty(blockImage) && displayName === 'Moniker')}
             center={center}
             color={color}
@@ -236,7 +239,12 @@ const Table = ({
               </Link>
             ) : (
               <>
-                {value}
+                {externalLink && (
+                  <a href={externalLink} target="_blank" rel="noreferrer">
+                    {value}
+                  </a>
+                )}
+                {!externalLink && value}
                 {icon && <Sprite icon={icon} size={size} color={iconColor} />}
                 {copy && <CopyValue value={raw} title={`Copy ${hover}`} />}
               </>